add new files

+++ src/light/event_loop_thread_pool.cpp
@@ -22,7 +22,7 @@ void EventLoopThreadPool::start(uint32_t threadCount) {
 
 	_started = true;
 
-	threadCount = std::max(threadCount, 1);
+	threadCount = std::max(threadCount, (uint32_t)1);
 	for (size_t i=0; i<threadCount; i++) {
 		boost::shared_ptr<EventLoopThread> threadPtr(new EventLoopThread((boost::format("%s:%d") % _name % (i+1)).str()));
 		threadPtr->start();
